fix balance amount type in transaction class

+++ src/main/java/com/novawallet/entity/Transaction.java
@@ -44,7 +44,7 @@ public class Transaction {
      *
      * @return the amount
      */
-    public double getAmount() {
+    public float getAmount() {
         return balance.getAmount();
     }
